chore: remove dead code and unused exports
- Remove TurnResult interface (never used)
- Remove rollEncounter function (forceEncounter used instead)
- Remove 9 unused item getter functions (getItemsByCategory, getAll*Items)
- Remove counterChance from DerivedBattleStats (calculated but never used)
- Remove applyStatModifier (only used in tests, moved to testing utils)
- Remove dotType return from processStatusEffects (never consumed)
- Remove ActionResult<T> interface (marked as future use, never used)
- Remove TICKS_PER_MONTH, ticksToMs (never imported)
- Remove createDefaultBattleStats from types (moved to testing utils)
- Make TICK_DURATION_SECONDS non-exported (only used internally)
- Make LEARNABLE_MOVES non-exported (only used to compose ALL_MOVES)

Based on code hygiene review in REVIEW.md
