+++ src/game/data/items/index.ts
@@ -2,17 +2,7 @@
  * Combined item database with lookup functions.
  */
 
-import type {
-  BattleItem,
-  CleaningItem,
-  DrinkItem,
-  EquipmentItem,
-  FoodItem,
-  Item,
-  MaterialItem,
-  MedicineItem,
-  ToyItem,
-} from "@/game/types/item";
+import type { Item } from "@/game/types/item";
 import { BATTLE_ITEMS, BATTLE_ITEMS_LIST } from "./battle";
 import { CLEANING_ITEMS, CLEANING_ITEMS_LIST } from "./cleaning";
 import { DRINK_ITEMS, DRINK_ITEMS_LIST } from "./drinks";
@@ -41,71 +31,6 @@ export const ALL_ITEMS: readonly Item[] = [
  */
 export function getItemById(id: string): Item | undefined {
   return ALL_ITEMS.find((item) => item.id === id);
-}
-
-/**
- * Get items by category.
- */
-export function getItemsByCategory<T extends Item>(
-  category: T["category"],
-): T[] {
-  return ALL_ITEMS.filter((item) => item.category === category) as T[];
-}
-
-/**
- * Get all food items.
- */
-export function getAllFoodItems(): FoodItem[] {
-  return [...FOOD_ITEMS_LIST];
-}
-
-/**
- * Get all drink items.
- */
-export function getAllDrinkItems(): DrinkItem[] {
-  return [...DRINK_ITEMS_LIST];
-}
-
-/**
- * Get all cleaning items.
- */
-export function getAllCleaningItems(): CleaningItem[] {
-  return [...CLEANING_ITEMS_LIST];
-}
-
-/**
- * Get all toy items.
- */
-export function getAllToyItems(): ToyItem[] {
-  return [...TOY_ITEMS_LIST];
-}
-
-/**
- * Get all medicine items.
- */
-export function getAllMedicineItems(): MedicineItem[] {
-  return [...MEDICINE_ITEMS_LIST];
-}
-
-/**
- * Get all battle items.
- */
-export function getAllBattleItems(): BattleItem[] {
-  return [...BATTLE_ITEMS_LIST];
-}
-
-/**
- * Get all equipment items.
- */
-export function getAllEquipmentItems(): EquipmentItem[] {
-  return [...EQUIPMENT_ITEMS_LIST];
-}
-
-/**
- * Get all material items.
- */
-export function getAllMaterialItems(): MaterialItem[] {
-  return [...MATERIAL_ITEMS_LIST];
 }
 
 // Re-export item objects (enum-like access via FOOD_ITEMS.KIBBLE.id)
